ef database

+++ Geolocation.Application/Geolocation.Application/Handlers/DeleteLocationCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Geolocation.Application.CQRS;
+using Geolocation.Application.Exceptions;
 using Geolocation.Core.Repository;
 using System.Threading.Tasks;
 
@@ -15,6 +16,12 @@ namespace Geolocation.Application.Handlers
 
         public async Task HandleAsync(DeleteLocation command)
         {
+            var exists = await _repository.ExistsAsync(command.LocationId);
+            if (!exists)
+            {
+                throw new GeoLocationNotFoundException(command.LocationId);
+            }
+
             await _repository.DeleteAsync(command.LocationId);
         }
     }
